requirements.txt
APScheduler==3.6.3
asn1crypto==1.5.1
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
blinker==1.4
Brotli==1.0.9
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
chart-studio==1.1.0
click==8.0.3
cloud-init==20.2
coincurve==15.0.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.6
coverage==6.5.0
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
fonttools==4.38.0
gcloud==0.18.3
google-auth==2.6.0
google-auth-oauthlib==0.5.0
googleapis-common-protos==1.53.0
green==3.3.0
gspread==5.2.0
httplib2==0.20.2
hyperlink==19.0.0
idna==3.3
importlib-metadata==1.5.0
incremental==16.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jws==0.1.3
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.9.1
macaroonbakery==1.3.1
MarkupSafe==2.0.1
matplotlib==3.6.2
more-itertools==4.2.0
mpl-finance==0.10.1
mplfinance==0.12.9b5
netifaces==0.10.4
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
plotly==5.11.0
protobuf==3.19.1
psycopg==3.1.4
psycopg2-binary==2.9.5
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycryptodome==3.11.0
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.6
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
pyTelegramBotAPI==4.8.0
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-jwt==2.0.1
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
retrying==1.3.3
rsa==4.9
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.10
systemd-python==234
TA-Lib==0.4.25
telegram==0.0.1
tenacity==8.1.0
tornado==6.1
Twisted==18.9.0
typing-extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.3.6
urllib3==1.25.8
wadllib==1.3.3
zipp==1.0.0
zope.interface==4.7.1
